+++ src/Cake.Git/GitAliases.Tags.cs
@@ -12,10 +12,15 @@ namespace Cake.Git
     partial class GitAliases
     {
         /// <summary>
-        /// Gets a list of all tags from the repo
+        /// Gets a list of all tags from the repository.
         /// </summary>
-        /// <param name="context"></param>
-        /// <param name="repositoryDirectoryPath"></param>
+        /// <remarks>
+        /// If you need to access the targets of the tags use <see cref="GitTags(ICakeContext, DirectoryPath, bool)"/>
+        /// to make sure targerts are loaded.
+        /// </remarks>
+        /// <param name="context">The context.</param>
+        /// <param name="repositoryDirectoryPath">Path to repository.</param>
+        /// <returns>List of all tags from the repository.</returns>
         /// <exception cref="ArgumentNullException"></exception>
         [CakeMethodAlias]
         [CakeAliasCategory("Tags")]
@@ -29,8 +34,49 @@ namespace Cake.Git
             if (repositoryDirectoryPath == null)
                 throw new ArgumentNullException(nameof(repositoryDirectoryPath));
 
+            return context.GitTags(repositoryDirectoryPath, false);
+        }
+
+        /// <summary>
+        /// Gets a list of all tags from the repository with the option to load targets of the tags.
+        /// </summary>
+        /// <remarks>
+        /// If you need to access the targets of the tags, set <paramref name="loadTargets"/> to <see langword="true"/>.
+        /// This will make sure that the targets are loaded before the <see cref="Repository"/> is disposed.
+        /// Otherwise, accessing a tag's target will throw an exception.
+        /// </remarks>
+        /// <param name="context">The context.</param>
+        /// <param name="repositoryDirectoryPath">Path to repository.</param>
+        /// <param name="loadTargets">A value indicating whether targets of the tags should be loaded.</param>
+        /// <returns>List of all tags from the repository.</returns>
+        /// <exception cref="ArgumentNullException"></exception>
+        [CakeMethodAlias]
+        [CakeAliasCategory("Tags")]
+        public static List<Tag> GitTags(
+            this ICakeContext context,
+            DirectoryPath repositoryDirectoryPath,
+            bool loadTargets)
+        {
+            if (context == null)
+                throw new ArgumentNullException(nameof(context));
+
+            if (repositoryDirectoryPath == null)
+                throw new ArgumentNullException(nameof(repositoryDirectoryPath));
+
             var retval =
-                context.UseRepository(repositoryDirectoryPath, repo => SortedTags(repo.Tags, t => t));
+                context.UseRepository(
+                    repositoryDirectoryPath,
+                    repo => SortedTags(
+                        repo.Tags,
+                        t =>
+                        {
+                            if (loadTargets)
+                            {
+                                _ = t.PeeledTarget;
+                            }
+
+                            return t;
+                        }));
 
             return retval;
         }